New upstream version 11.0.0~b2

+++ neutron/plugins/ml2/drivers/type_vlan.py
@@ -15,25 +15,22 @@
 
 import sys
 
+from neutron_lib import context
 from neutron_lib import exceptions as exc
+from neutron_lib.plugins.ml2 import api
 from oslo_config import cfg
 from oslo_log import log
 from six import moves
 
 from neutron._i18n import _, _LE, _LI, _LW
-from neutron.common import _deprecate
 from neutron.conf.plugins.ml2.drivers import driver_type
-from neutron import context
 from neutron.db import api as db_api
 from neutron.db.models.plugins.ml2 import vlanallocation as vlan_alloc_model
 from neutron.plugins.common import constants as p_const
 from neutron.plugins.common import utils as plugin_utils
-from neutron.plugins.ml2 import driver_api as api
 from neutron.plugins.ml2.drivers import helpers
 
 LOG = log.getLogger(__name__)
-
-_deprecate._moved_global('VlanAllocation', new_module=vlan_alloc_model)
 
 driver_type.register_ml2_drivers_vlan_opts()
 
@@ -69,8 +66,7 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
         with db_api.context_manager.writer.using(ctx):
             # get existing allocations for all physical networks
             allocations = dict()
-            allocs = (ctx.session.query(vlan_alloc_model.VlanAllocation).
-                      with_lockmode('update'))
+            allocs = ctx.session.query(vlan_alloc_model.VlanAllocation)
             for alloc in allocs:
                 if alloc.physical_network not in allocations:
                     allocations[alloc.physical_network] = set()
@@ -102,7 +98,19 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
                                           {'vlan_id': alloc.vlan_id,
                                            'physical_network':
                                            physical_network})
-                                ctx.session.delete(alloc)
+                                # This UPDATE WHERE statement blocks anyone
+                                # from concurrently changing the allocation
+                                # values to True while our transaction is
+                                # open so we don't accidentally delete
+                                # allocated segments. If someone has already
+                                # allocated, count will return 0 so we don't
+                                # delete.
+                                count = allocs.filter_by(
+                                    allocated=False, vlan_id=alloc.vlan_id,
+                                    physical_network=physical_network
+                                ).update({"allocated": False})
+                                if count:
+                                    ctx.session.delete(alloc)
                     del allocations[physical_network]
 
                 # add missing allocatable vlans to table
@@ -141,7 +149,7 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
         if physical_network:
             if physical_network not in self.network_vlan_ranges:
                 msg = (_("physical_network '%s' unknown "
-                         " for VLAN provider network") % physical_network)
+                         "for VLAN provider network") % physical_network)
                 raise exc.InvalidInput(error_message=msg)
             if segmentation_id:
                 if not plugin_utils.is_valid_vlan_tag(segmentation_id):
@@ -149,6 +157,12 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
                              "%(max)s)") %
                            {'min': p_const.MIN_VLAN_TAG,
                             'max': p_const.MAX_VLAN_TAG})
+                    raise exc.InvalidInput(error_message=msg)
+            else:
+                if not self.network_vlan_ranges.get(physical_network):
+                    msg = (_("Physical network %s requires segmentation_id "
+                             "to be specified when creating a provider "
+                             "network") % physical_network)
                     raise exc.InvalidInput(error_message=msg)
         elif segmentation_id:
             msg = _("segmentation_id requires physical_network for VLAN "
@@ -188,8 +202,12 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
                 api.MTU: self.get_mtu(alloc.physical_network)}
 
     def allocate_tenant_segment(self, context):
-        alloc = self.allocate_partially_specified_segment(context)
-        if not alloc:
+        for physnet in self.network_vlan_ranges:
+            alloc = self.allocate_partially_specified_segment(
+                context, physical_network=physnet)
+            if alloc:
+                break
+        else:
             return
         return {api.NETWORK_TYPE: p_const.TYPE_VLAN,
                 api.PHYSICAL_NETWORK: alloc.physical_network,
@@ -203,7 +221,7 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
         ranges = self.network_vlan_ranges.get(physical_network, [])
         inside = any(lo <= vlan_id <= hi for lo, hi in ranges)
 
-        with context.session.begin(subtransactions=True):
+        with db_api.context_manager.writer.using(context):
             query = (context.session.query(vlan_alloc_model.VlanAllocation).
                      filter_by(physical_network=physical_network,
                                vlan_id=vlan_id))
@@ -236,6 +254,3 @@ class VlanTypeDriver(helpers.SegmentTypeDriver):
         if physical_network in self.physnet_mtus:
             mtu.append(int(self.physnet_mtus[physical_network]))
         return min(mtu) if mtu else 0
-
-
-_deprecate._MovedGlobals()
